electron: rename shared/auth.ts to shared/electronAuth.ts
