+++ packages/shared/src/index.ts
@@ -39,4 +39,4 @@ export * as tiptap from './logic/tiptap';
 export * as utilHooks from './logic/utilHooks';
 export * from './debug';
 export * from './perf';
-export * from './auth';
+export * from './electrtonAuth';
